An MNIST classifier using Keras
MNIST is a simple computer vision dataset. It consists of images of handwritten digits between zero and nine. So there are only ten possible things that a given image can be. 
I have  implemented  a small subsection of object recognition—digit recognition. Using Keras ,
I take hand-drawn images of the numbers 0-9 and build and train a neural network to recognize and predict the correct label for the digit displayed.